split up the modules on responsibilities
SRP and provide a mechanism for comparability
