+++ app/operations/concerns/content_from_params.rb
@@ -1,0 +1,14 @@
+module ContentFromParams
+  def content_from_params(ps, content_fields)
+    yield ps if block_given?
+    content = ps.slice(*content_fields)
+    content[:language] = ps[:primary_language]
+    if ps.has_key? :urls
+      urls, labels = extract_url_labels(ps[:urls])
+      content[:url_labels] = labels
+      ps[:urls] = urls
+    end
+    ps.except!(*content_fields)
+    content.select { |k,v| !!v }
+  end
+end
