+++ app/operations/organizations/create.rb
@@ -1,6 +1,7 @@
 module Organizations
   class Create < Operation
     include UrlLabels
+    include ContentFromParams
 
     string :name
     string :display_name
